Boot & web socket config splitted
