+++ spring-boot-samples/spring-boot-sample-websocket-tomcat/src/main/java/samples/websocket/tomcat/SampleTomcatWebSocketApplication.java
@@ -16,73 +16,11 @@
 
 package samples.websocket.tomcat;
 
-import samples.websocket.tomcat.client.GreetingService;
-import samples.websocket.tomcat.client.SimpleGreetingService;
-import samples.websocket.tomcat.echo.DefaultEchoService;
-import samples.websocket.tomcat.echo.EchoService;
-import samples.websocket.tomcat.echo.EchoWebSocketHandler;
-import samples.websocket.tomcat.reverse.ReverseWebSocketEndpoint;
-import samples.websocket.tomcat.snake.SnakeWebSocketHandler;
-
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.socket.WebSocketHandler;
-import org.springframework.web.socket.config.annotation.EnableWebSocket;
-import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
-import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
-import org.springframework.web.socket.handler.PerConnectionWebSocketHandler;
-import org.springframework.web.socket.server.standard.ServerEndpointExporter;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@Configuration
-@EnableAutoConfiguration
-@EnableWebSocket
-public class SampleTomcatWebSocketApplication extends SpringBootServletInitializer
-		implements WebSocketConfigurer {
-
-	@Override
-	public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-		registry.addHandler(echoWebSocketHandler(), "/echo").withSockJS();
-		registry.addHandler(snakeWebSocketHandler(), "/snake").withSockJS();
-	}
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(SampleTomcatWebSocketApplication.class);
-	}
-
-	@Bean
-	public EchoService echoService() {
-		return new DefaultEchoService("Did you say \"%s\"?");
-	}
-
-	@Bean
-	public GreetingService greetingService() {
-		return new SimpleGreetingService();
-	}
-
-	@Bean
-	public WebSocketHandler echoWebSocketHandler() {
-		return new EchoWebSocketHandler(echoService());
-	}
-
-	@Bean
-	public WebSocketHandler snakeWebSocketHandler() {
-		return new PerConnectionWebSocketHandler(SnakeWebSocketHandler.class);
-	}
-
-	@Bean
-	public ReverseWebSocketEndpoint reverseWebSocketEndpoint() {
-		return new ReverseWebSocketEndpoint();
-	}
-
-	@Bean
-	public ServerEndpointExporter serverEndpointExporter() {
-		return new ServerEndpointExporter();
-	}
+@SpringBootApplication
+public class SampleTomcatWebSocketApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(SampleTomcatWebSocketApplication.class, args);
