+++ spring-boot-samples/spring-boot-sample-websocket-tomcat/src/main/java/samples/websocket/tomcat/echo/EchoWebSocketHandler.java
@@ -16,39 +16,36 @@
 
 package samples.websocket.tomcat.echo;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
-import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.WebSocketSession;
 import org.springframework.web.socket.handler.TextWebSocketHandler;
 
-/**
- * Echo messages by implementing a Spring {@link WebSocketHandler} abstraction.
- */
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 public class EchoWebSocketHandler extends TextWebSocketHandler {
 
-	private static Logger logger = LoggerFactory.getLogger(EchoWebSocketHandler.class);
-
 	private final EchoService echoService;
-
-	public EchoWebSocketHandler(EchoService echoService) {
-		this.echoService = echoService;
+	
+	public EchoWebSocketHandler(EchoService echoService){
+		this.echoService=echoService;
 	}
 
 	@Override
 	public void afterConnectionEstablished(WebSocketSession session) {
-		logger.debug("Opened new session in instance " + this);
+		log.info("Opened new session in instance " + this);
 	}
 
 	@Override
 	public void handleTextMessage(WebSocketSession session, TextMessage message)
 			throws Exception {
 		String echoMessage = this.echoService.getMessage(message.getPayload());
-		logger.debug(echoMessage);
-		session.sendMessage(new TextMessage(echoMessage));
+		for(int i=0; i<3; i++){
+			log.info(echoMessage);
+			session.sendMessage(new TextMessage(echoMessage));
+		}
 	}
 
 	@Override
